# api/src/main/resources/moleskine-openapi-spec.yaml
openapi: 3.0.0
info:
    title: Todo Application API
    version: 1.0.0-SNAPSHOT
    description: A simple API for managing todos
    contact:
        email: yev@devatlant.com
        name: Yevgen Voronetski
        url: https://devatlant.com

servers:
    - url: http://localhost:8080

components:
    schemas:
        ToDoDto:
            type: object
            required:
                - title
            properties:
                id:
                    type: integer
                    description: technical id. Bad idea to expose but for my learning purposes it's ok.
                    example: 123
                title:
                    type: string
                    description: title
                    example: "my title"
                isCompleted:
                    type: boolean
                    description: is this todo item is done
                    example: true
                    default: false
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
        ApiErrorItem:
            type: object
            required:
                - name
                - errorDescription
                - gitId
            properties:
                name:
                    type: string
                errorDescription:
                    type: string
                gitId:
                    type: string

        ApiError:
            type: array
            items:
                $ref: "#/components/schemas/ApiErrorItem"


paths:
    /todos:
        post:
            summary: Create a new todo
            operationId: createTodo
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ToDoDto"
            responses:
                '201':
                    description: Todo created successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ToDoDto"
                '400':
                    description: Invalid input
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ApiError"
        patch:
            summary: update existing todo
            operationId: updateTodo
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/ToDoDto"
            responses:
                201:
                    description: todo was updated successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ToDoDto"
                400:
                    description: Invalid input
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ApiError"
                404:
                    description: entity not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ApiError"
                500:
                    description: internal server error
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ApiError"
    /todos/{id}:
        get:
            summary: Retrieve a todo by ID
            operationId: getTodoById
            parameters:
                - name: id
                  in: path
                  required: true
                  description: The ID of the todo to retrieve
                  schema:
                      type: integer
            responses:
                '200':
                    description: Todo retrieved successfully
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ToDoDto"
                '404':
                    description: Todo not found

    /todos/search:
        get:
            summary: Search todos by title
            operationId: searchTodosByTitle
            parameters:
                - name: title
                  in: query
                  required: true
                  description: The title of the todo to search for
                  schema:
                      type: string
            responses:
                '200':
                    description: List of todos matching the search criteria
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/ToDoDto'
    /todos/:
        get:
            summary: get all todos from the database
            operationId: findAll
            responses:
                '200':
                    description: List of todos matching the search criteria
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/ToDoDto'